+++ packages/core/postgrest-js/test/relationships.test.ts
@@ -5,7 +5,7 @@ import { z } from 'zod'
 import { Json } from '../src/select-query-parser/types'
 import { RequiredDeep } from 'type-fest'
 
-const REST_URL = 'http://localhost:3000'
+const REST_URL = 'http://localhost:54321/rest/v1'
 const postgrest = new PostgrestClient<Database>(REST_URL)
 
 const UsersRowSchema = z.object({
@@ -42,7 +42,7 @@ test('nested query with selective fields', async () => {
           },
           {
             "id": 4,
-            "message": "Some message on channel wihtout details",
+            "message": "Some message on channel without details",
           },
         ],
         "username": "supabot",
@@ -100,7 +100,7 @@ test('nested query with multiple levels and selective fields', async () => {
               "slug": "other",
             },
             "id": 4,
-            "message": "Some message on channel wihtout details",
+            "message": "Some message on channel without details",
           },
         ],
         "username": "supabot",
@@ -226,7 +226,7 @@ test('one-to-many relationship', async () => {
             "channel_id": 3,
             "data": null,
             "id": 4,
-            "message": "Some message on channel wihtout details",
+            "message": "Some message on channel without details",
             "username": "supabot",
           },
         ],