+++ packages/core/realtime-js/src/RealtimeChannel.ts
@@ -545,21 +545,22 @@ export default class RealtimeChannel {
     payload: any,
     opts: { timeout?: number } = {}
   ): Promise<{ success: true } | { success: false; status: number; error: string }> {
-    const authorization = this.socket.accessTokenValue
-      ? `Bearer ${this.socket.accessTokenValue}`
-      : ''
-
     if (payload === undefined || payload === null) {
       return Promise.reject('Payload is required for httpSend()')
     }
 
+    const headers: Record<string, string> = {
+      apikey: this.socket.apiKey ? this.socket.apiKey : '',
+      'Content-Type': 'application/json',
+    }
+
+    if (this.socket.accessTokenValue) {
+      headers['Authorization'] = `Bearer ${this.socket.accessTokenValue}`
+    }
+
     const options = {
       method: 'POST',
-      headers: {
-        Authorization: authorization,
-        apikey: this.socket.apiKey ? this.socket.apiKey : '',
-        'Content-Type': 'application/json',
-      },
+      headers,
       body: JSON.stringify({
         messages: [
           {
@@ -617,16 +618,18 @@ export default class RealtimeChannel {
       )
 
       const { event, payload: endpoint_payload } = args
-      const authorization = this.socket.accessTokenValue
-        ? `Bearer ${this.socket.accessTokenValue}`
-        : ''
+      const headers: Record<string, string> = {
+        apikey: this.socket.apiKey ? this.socket.apiKey : '',
+        'Content-Type': 'application/json',
+      }
+
+      if (this.socket.accessTokenValue) {
+        headers['Authorization'] = `Bearer ${this.socket.accessTokenValue}`
+      }
+
       const options = {
         method: 'POST',
-        headers: {
-          Authorization: authorization,
-          apikey: this.socket.apiKey ? this.socket.apiKey : '',
-          'Content-Type': 'application/json',
-        },
+        headers,
         body: JSON.stringify({
           messages: [
             {
@@ -772,7 +775,8 @@ export default class RealtimeChannel {
             bindId &&
             payload.ids?.includes(bindId) &&
             (bindEvent === '*' ||
-              bindEvent?.toLocaleLowerCase() === payload.data?.type.toLocaleLowerCase())
+              bindEvent?.toLocaleLowerCase() === payload.data?.type.toLocaleLowerCase()) &&
+            (!bind.filter?.table || bind.filter.table === payload.data?.table)
           )
         } else {
           const bindEvent = bind?.filter?.event?.toLocaleLowerCase()