+++ packages/core/storage-js/jest.config.js
@@ -2,7 +2,7 @@ module.exports = {
   preset: 'ts-jest',
   testEnvironment: 'node',
   collectCoverage: true,
-  coverageDirectory: './test/coverage',
+  coverageDirectory: './coverage',
   coverageReporters: ['lcov'],
   collectCoverageFrom: [
     './src/**/*.{js,ts}',